% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{prov.analyze.run.addin}
\alias{prov.analyze.run.addin}
\title{Addin for RStudio}
\usage{
prov.analyze.run.addin()
}
\description{
Runs the analyzer on the active script. This function is added as an item in
the Addins menu.
}
