% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provAnalyzeR.R
\name{prov.analyze.run}
\alias{prov.analyze.run}
\title{prov.analyze.run}
\usage{
prov.analyze.run(r.script, save = FALSE, create.zip = FALSE, ...)
}
\arguments{
\item{r.script}{the name of a file containing an R script}

\item{...}{extra parameters are passed to the provenance collector.  See rdt's prov.run function
or rdtLites's prov.run function for details.}
}
\description{
prov.analyze.run executes a script, collects provenance, and outputs a
text summary of any coding anomalies to the console.
}
\examples{
\dontrun{
ADD EXAMPLE}
}
